--- a/django/contrib/gis/db/models/fields.py
+++ b/django/contrib/gis/db/models/fields.py
@@ -5,6 +5,10 @@ from django.contrib.gis import forms, gdal
 from django.contrib.gis.db.models.proxy import SpatialProxy
 from django.contrib.gis.gdal.error import GDALException
 from django.contrib.gis.geometry.backend import Geometry, GeometryException
+from django.contrib.gis.geos import (
+    GeometryCollection, LineString, MultiLineString, MultiPoint, MultiPolygon,
+    Point, Polygon,
+)
 from django.core.exceptions import ImproperlyConfigured
 from django.db.models.fields import Field
 from django.utils.translation import gettext_lazy as _
@@ -196,6 +200,7 @@ class GeometryField(BaseSpatialField):
     form_class = forms.GeometryField
     # The OpenGIS Geometry name.
     geom_type = 'GEOMETRY'
+    geom_class = None
 
     def __init__(self, verbose_name=None, dim=2, geography=False, *, extent=(-180.0, -90.0, 180.0, 90.0),
                  tolerance=0.05, **kwargs):
@@ -268,42 +273,49 @@ class GeometryField(BaseSpatialField):
 # The OpenGIS Geometry Type Fields
 class PointField(GeometryField):
     geom_type = 'POINT'
+    geom_class = Point
     form_class = forms.PointField
     description = _("Point")
 
 
 class LineStringField(GeometryField):
     geom_type = 'LINESTRING'
+    geom_class = LineString
     form_class = forms.LineStringField
     description = _("Line string")
 
 
 class PolygonField(GeometryField):
     geom_type = 'POLYGON'
+    geom_class = Polygon
     form_class = forms.PolygonField
     description = _("Polygon")
 
 
 class MultiPointField(GeometryField):
     geom_type = 'MULTIPOINT'
+    geom_class = MultiPoint
     form_class = forms.MultiPointField
     description = _("Multi-point")
 
 
 class MultiLineStringField(GeometryField):
     geom_type = 'MULTILINESTRING'
+    geom_class = MultiLineString
     form_class = forms.MultiLineStringField
     description = _("Multi-line string")
 
 
 class MultiPolygonField(GeometryField):
     geom_type = 'MULTIPOLYGON'
+    geom_class = MultiPolygon
     form_class = forms.MultiPolygonField
     description = _("Multi polygon")
 
 
 class GeometryCollectionField(GeometryField):
     geom_type = 'GEOMETRYCOLLECTION'
+    geom_class = GeometryCollection
     form_class = forms.GeometryCollectionField
     description = _("Geometry collection")