https://code.djangoproject.com/ticket/28518

Before:
```
In [3]: %timeit for x in City.objects.values_list('point')[:1000]: pass
49.5 ms ± 367 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
After:
```
In [4]: %timeit for x in City.objects.values_list('point')[:1000]: pass
32.1 ms ± 1.54 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```